Mapents & scripting changes + some fixes

+++ src/client/game/scripting/animation.cpp
@@ -5,7 +5,7 @@
 
 namespace scripting
 {
-	animation::animation(unsigned int value)
+	animation::animation(uint64_t value)
 		: value_(value)
 	{
 	}
